refactor: add Header in common components
