+++ src/pages/addCard/components/index.ts
@@ -5,4 +5,3 @@ export { default as CardNumberInput } from './addCard/CardNumberInput';
 export { default as CardPasswordInput } from './addCard/CardPasswordInput';
 export { default as CardSecurityCodeInput } from './addCard/CardSecurityCodeInput';
 export { default as Dot } from './addCard/Dot';
-export { default as Header } from './Header';
